finished the puzzle board
