+++ helper_function.cpp
