+++ helper_functions.hpp
@@ -3,8 +3,10 @@
 
 #include <string>
 #include "constants/constants.hpp"
+#include "helping_structure.hpp"
 
 std::string decode_date(std::string boardRep);
 unsigned int date_to_index(std::string input);
+
 
 #endif // !HELPER_FUNCTION_HPP_
