+++ helping_structure.hpp
@@ -1,6 +1,7 @@
 #ifndef HELPING_STRUCTURE_HPP_
 #define HELPING_STRUCTURE_HPP_
 
+
 struct coord
 {
     unsigned int row:4;
