+++ main.cpp
@@ -8,25 +8,14 @@
 #include <stdlib.h>
 
 #include "puzzle_piece/puzzle_piece.hpp"
+#include "puzzle_board/puzzle_board.hpp"
 #include "constants/constants.hpp"
 #include "helper_functions.hpp"
+#include "helping_structure.hpp"
 
 #define MAX_LOCATION 38
 #define DEBUG   0
-
-
-
-class puzzle_board
-{
-    private:
-    std::string m_board[7];
-    public:
-    puzzle_board();
-    bool try_to_fit(puzzle_piece pieceToFit , coord pieceLocation);
-    void remove_shape(puzzle_piece pieceToRem , coord pieceLocation);
-    void reset_board();
-    std::string board_to_string();
-};
+#define MAX_TIME 1*60*CLOCKS_PER_SEC
 
 inline unsigned int coord_to_loc(coord in);
 inline coord loc_to_coord(unsigned int in);
@@ -37,7 +26,6 @@ struct backtrack
     puzzle_piece piece;
 };
 
-#define MAX_TIME 1*60*CLOCKS_PER_SEC
 
 int main (int argc, char * argv[])
 {
@@ -241,76 +229,6 @@ int main (int argc, char * argv[])
     return 0;
 }
 
-
-puzzle_board::puzzle_board()
-{
-    this->reset_board();
-}
-
-bool puzzle_board::try_to_fit(puzzle_piece pieceToFit , coord pieceLocation)
-{
-    char toFill = pieceToFit.get_pieceName();
-    coord parts[6];
-    unsigned int i = 0;
-    bool canFit = true;
-    while(canFit && pieceToFit.get_next_coord(parts[i]))
-    {
-        parts[i].row+= pieceLocation.row;
-        parts[i].col+= pieceLocation.col;
-        if( parts[i].row > 6 || parts[i].col > 6 || this->m_board[parts[i].row][parts[i].col]!='.')
-        {
-            canFit=false;
-        }
-        i++;
-    }
-    if(canFit)
-    {
-        while(i>0)
-        {
-            i--;
-            this->m_board[parts[i].row][parts[i].col] = toFill;
-        }
-    }
-    return canFit;
-}
-
-void puzzle_board::remove_shape(puzzle_piece pieceToRem , coord pieceLocation)
-{
-    char pieceName = pieceToRem.get_pieceName();
-    coord tempCoord = {0,0};
-    unsigned int i = 0;
-    bool canFit = true;
-    while(pieceToRem.get_next_coord(tempCoord))
-    {
-        tempCoord.row+= pieceLocation.row;
-        tempCoord.col+= pieceLocation.col;
-        if( tempCoord.row < 7 && tempCoord.col < 7 && pieceName == this->m_board[tempCoord.row][tempCoord.col])
-        {
-           this->m_board[tempCoord.row][tempCoord.col] ='.';
-        }
-    }
-}
-
-void puzzle_board::reset_board()
-{
-    this->m_board[0] = "...... ";
-    this->m_board[1] = "...... ";
-    this->m_board[2] = ".......";
-    this->m_board[3] = ".......";
-    this->m_board[4] = ".......";
-    this->m_board[5] = ".......";
-    this->m_board[6] = "...    ";
-}
-
-std::string puzzle_board::board_to_string()
-{
-    std::string result = "";
-    for (int i = 0; i < 7; i++)
-    {
-        result+=this->m_board[i]+"\n";
-    }
-    return result;
-}
 
 inline unsigned int coord_to_loc(coord in)
 {
